Minimal API
Minimal API enables you to create an API with only a few lines of code. It has all the major features that you're used to like dependency injection, talking to databases, and route management. A minimal API differs from a controller-based API because you explicitly specify the routes you need instead of relying on a convention-based approach like with a controller-based API